+++ corbeans-corda-webserver/src/test/kotlin/com/github/manosbatsis/corbeans/corda/webserver/CorbeansSpringExtensionIntegrationTest.kt
@@ -29,6 +29,7 @@ import com.github.manosbatsis.corbeans.test.integration.WithImplicitNetworkIT
 import net.corda.core.identity.CordaX500Name
 import net.corda.core.identity.Party
 import net.corda.core.utilities.NetworkHostAndPort
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Test
@@ -40,6 +41,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMock
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.boot.test.web.client.TestRestTemplate
 import org.springframework.http.HttpHeaders
+import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.mock.web.MockMultipartFile
 import org.springframework.test.web.servlet.MockMvc
@@ -93,7 +95,7 @@ class CorbeansSpringExtensionIntegrationTest {
     @Autowired
     lateinit var restTemplate: TestRestTemplate
 
-    //@Test
+    @Test
     fun `Can use both default node and multiple node controller endpoints`() {
         val defaultNodeMe = this.restTemplate.getForObject("/api/node/me", Map::class.java)
         assertEquals("me", defaultNodeMe.keys.first())
@@ -102,7 +104,7 @@ class CorbeansSpringExtensionIntegrationTest {
     }
 
 
-    //@Test
+    @Test
     fun `Can inject services`() {
         logger.info("services: {}", services)
         assertNotNull(this.networkService)
@@ -111,36 +113,36 @@ class CorbeansSpringExtensionIntegrationTest {
         assertTrue(this.services.keys.isNotEmpty())
     }
 
-    //@Test
+    @Test
     fun `Can inject custom service`() {
         logger.info("customCervice: {}", customCervice)
         assertNotNull(this.customCervice)
         assertTrue(this.customCervice.dummy())
     }
 
-    //@Test
+    @Test
     fun `Can retrieve node identity`() {
         assertNotNull(service.myIdentity)
     }
 
-    //@Test
+    @Test
     fun `Can retrieve peer identities`() {
         assertNotNull(service.identities())
     }
 
-    //@Test
+    @Test
     fun `Can retrieve notaries`() {
         val notaries: List<Party> = service.notaries()
         assertNotNull(notaries)
     }
 
-    //@Test
+    @Test
     fun `Can retrieve flows`() {
         val flows: List<String> = service.flows()
         assertNotNull(flows)
     }
 
-    //@Test
+    @Test
     fun `Can retrieve addresses`() {
         val addresses: List<NetworkHostAndPort> = service.addresses()
         assertNotNull(addresses)
@@ -162,7 +164,7 @@ class CorbeansSpringExtensionIntegrationTest {
     }
 
 
-    //@Test
+    @Test
     @Throws(Exception::class)
     fun `Can save and retrieve regular files as attachments`() {
         val headers = HttpHeaders()
@@ -194,12 +196,60 @@ class CorbeansSpringExtensionIntegrationTest {
         assertTrue(paths.containsAll(listOf("test.txt", "test.png")))
     }
 
-    //@Test
+    @Test
     @Throws(Exception::class)
     fun `Can save and retrieve single zip and jar files as attachments`() {
         testArchiveUploadAndDownload("test.zip", "application/zip")
         testArchiveUploadAndDownload("test.jar", "application/java-archive")
     }
+
+    @Test
+    fun `Can access swagger UI`() {
+        // Check swagger endpoint
+        this.mockMvc.perform(get("/api-docs"))
+                .andExpect(status().isOk())
+        // Check Swagger UI
+        this.mockMvc.perform(get("/swagger-ui.html"))
+                .andExpect(status().isOk())
+    }
+
+    @Test
+    fun `Can see Corda details within Actuator info endpoint response`() {
+        logger.debug("testInfoContributor, called")
+        val entity = this.restTemplate
+                .getForEntity("/actuator/info", Map::class.java)
+        // Ensure a 200 OK
+        Assertions.assertEquals(HttpStatus.OK, entity.statusCode)
+
+        val body = entity.body
+        Assertions.assertNotNull(body, "Actuator info must not be null")
+        val corda = body!!["corda"] as Map<*, *>?
+        // Validate corda information
+        validateCordaInfo(corda)
+    }
+
+    @Test
+    fun `Can access Corda custom Actuator endpoint`() {
+        logger.debug("testCordaEndpoint, called")
+        val serviceKeys = this.networkService.nodeServices.keys
+        logger.debug("testCordaEndpoint, serviceKeys: {}", serviceKeys)
+        val entity = this.restTemplate
+                .getForEntity("/actuator/corda", Map::class.java)
+        // Ensure a 200 OK
+        Assertions.assertEquals(HttpStatus.OK, entity.statusCode)
+        val corda = entity.body
+        // Validate corda information
+        validateCordaInfo(corda)
+    }
+
+
+    private fun validateCordaInfo(corda: Map<*, *>?) {
+        Assertions.assertNotNull(corda, "Actuator corda info must not be null")
+        val cordaNodes = corda!!["nodes"] as Map<String, Any>
+        Assertions.assertNotNull(cordaNodes["partyA"])
+        Assertions.assertNotNull(cordaNodes["partyB"])
+    }
+
 
     private fun uploadAttachmentFiles(vararg  file: MockMultipartFile): AttachmentReceipt {
         var attachmentReceipt: AttachmentReceipt? = null